requirements.txt
# API Framework
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6

# LLM Integration
transformers==4.33.2
torch==2.0.1
sentence-transformers==2.2.2

# Memory Store
redis>=4.6.0
# Alternatively: SQLite is part of Python's standard library

# PDF Processing
PyPDF2>=3.0.1
# Alternative: apache-tika>=2.6.0

# Email Processing
email-validator>=2.0.0

# JSON Processing
jsonschema>=4.17.3

# Utility Libraries
pydantic>=2.0.3
python-dotenv>=1.0.0

# For Bonus UI
jinja2>=3.1.2

# For Testing
pytest>=7.4.0
faker>=18.13.0
